experimentally setting hash values to 64bit

+++ src/types.h
@@ -484,8 +484,8 @@ typedef uint32_t nelts_t;
 typedef uint32_t ht_size_t;
 
 /* hash table entry size */
-typedef uint32_t hash_t;
-/* typedef uint64_t hash_t; */
+/* typedef uint32_t hash_t; */
+typedef uint64_t hash_t;
 /* typedef unsigned long hash_t; */
 
 /* degree size */
